pyproject.toml
[project]
name = "ecaa"
version = "0.2.0"
description = "Engineering Cognitive AI Agents"
authors = [{ name = "Donald Thompson", email = "witt3rd@witt3rd.com" }]
dependencies = [
    "chainlit>=1.1.301",
    "openai>=1.34.0",
    "litellm>=1.40.14",
    "python-frontmatter>=1.1.0",
    "loguru>=0.7.2",
    "chromadb>=0.5.20",
    "sentence-transformers>=3.0.1",
    "pytest-asyncio>=0.24.0",
    "newspaper3k>=0.2.8",
    "lxml[html_clean]>=5.3.0",
    "pydantic===2.10.1",            # Bug in Chainlit requires this version
    "function-schema>=0.4.4",
    "matplotlib>=3.9.2",
    "networkx>=3.4.2",
]
readme = "README.md"
requires-python = ">=3.12.4"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=3.7.1",
    "ipykernel>=6.29.5",
    "pytest>=8.3.2",
    "ruff>=0.6.5",
    "pip>=24.3.1",
]

[tool.ruff]
line-length = 55 # Book formatting
fix = true

[tool.basedpyright]
typeCheckingMode = "strict"
verboseOutput = true
venvPath = ".venv"
pythonVersion = "3.12"

[tool.hatch.build.targets.wheel]
packages = ["src", "examples"]

[tool.rye.scripts]
precommit = "pre-commit run --all-files"
ch02_chat = "chainlit run examples/ch02/winston_chat.py"
ch02_tool = "chainlit run examples/ch02/winston_tool.py"
ch04_zettelkasten = "chainlit run examples/ch04/winston_zettelkasten.py"
